requirements.txt
pillow==10.2.0
python-pptx==0.6.23